Close Goal form supports subscriptions
